requirements.txt
Flask

google-cloud-storage
selenium
tenacity
python-dotenv==1.0.1
fastapi==0.109.1
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
openai==1.11.1
pytest==8.0.0
httpx==0.26.0
python-multipart==0.0.6
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
pydantic==2.5.3